core/requirements.txt
amqp==5.3.1
python-dotenv==1.1.1
asgiref==3.9.1
async-timeout==5.0.1
audioread==3.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.5
decorator==5.2.1
Django==4.2.24
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
filelock==3.19.1
fsspec==2025.7.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
joblib==1.5.2
kombu==5.5.4
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.2.1
numba==0.60.0
numpy==1.24.3
openai-whisper==20250625
packaging==25.0
platformdirs==4.4.0
pooch==1.8.2
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pycparser==2.23
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
regex==2025.7.34
requests==2.32.4
scikit-learn==1.6.1
scipy==1.11.4
six==1.17.0
soundfile==0.12.1
soxr==1.0.0
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.11.0
torch==2.1.2
torchaudio==2.1.2
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
python-docx
reportlab